+++ SpendLess/SpendLess.Transactions/Services/ITransactionService.cs
@@ -7,16 +7,16 @@ namespace SpendLess.Transactions.Services
     {
         Task<List<bool>> CheckIfTransactionsHaveBeenImported(List<TransactionDto> transactions);
         Task<List<bool>> CheckIfTransactionsHaveBeenImported(List<List<string>> sourceDatas);
-        Task<List<int>> CreateTransactions(List<TransactionDto> transactions);
+        Task<List<long>> CreateTransactions(List<TransactionDto> transactions);
         Task<(Dictionary<string, decimal> BySource, Dictionary<string, decimal> ByDestination)> GetAmounts(List<TransactionFilter> filters);
         Task<List<string>> GetCategories();
-        Task<Dictionary<int, ExtendedTransactionDto>> GetPagedExtendedTransactions(List<TransactionFilter> filters, int pageSize, int page);
-        Task<Dictionary<int, TransactionDto>> GetPagedTransactions(List<TransactionFilter> filters, int pageSize, int page);
+        Task<Dictionary<long, ExtendedTransactionDto>> GetPagedExtendedTransactions(List<TransactionFilter> filters, long pageSize, long page);
+        Task<Dictionary<long, TransactionDto>> GetPagedTransactions(List<TransactionFilter> filters, long pageSize, long page);
         Task<List<string>> GetTags();
-        Task<Dictionary<int, TransactionDto>> GetTransactions(List<TransactionFilter> filters);
-        Task<int> DeleteTransactions(List<int> keys);
-        Task<bool> DeleteTransaction(int key);
-        Task<int> GetTransactionsCount(List<TransactionFilter> filters);
+        Task<Dictionary<long, TransactionDto>> GetTransactions(List<TransactionFilter> filters);
+        Task<int> DeleteTransactions(List<long> keys);
+        Task<bool> DeleteTransaction(long key);
+        Task<long> GetTransactionsCount(List<TransactionFilter> filters);
         Task<int> DeleteTransactions(List<TransactionFilter> filters);
     }
 }
