+++ SpendLess/SpendLess.Transactions/Storages/ITransactionStorage.cs
@@ -6,17 +6,17 @@ namespace SpendLess.Transactions.Storages
 {
     public interface ITransactionStorage
     {
-        Task<List<int>> AddTransactions(List<TransactionDto> transaction);
-        Task<TransactionDto> GetTransaction(int key);
-        (StorageOperation Operation, Func<List<int>> GetResult) CreateAddTransactionsOperation(List<TransactionDto> transactions);
+        Task<List<long>> AddTransactions(List<TransactionDto> transaction);
+        Task<TransactionDto> GetTransaction(long key);
+        (StorageOperation Operation, Func<List<long>> GetResult) CreateAddTransactionsOperation(List<TransactionDto> transactions);
         Task<List<bool>> CheckIfHasSourceDataHash(IEnumerable<long> hashes);
-        Task<Dictionary<int, TransactionDto>> GetTransactions(List<TransactionFilter> filters, int? limit = null, int? offset = null);
-        Task<int> GetTransactionsCount(List<TransactionFilter> filters);
+        Task<Dictionary<long, TransactionDto>> GetTransactions(List<TransactionFilter> filters, long? limit = null, long? offset = null);
+        Task<long> GetTransactionsCount(List<TransactionFilter> filters);
         Task<(Dictionary<string, decimal> BySource, Dictionary<string, decimal> ByDestination)> GetAmounts(List<TransactionFilter> filters);
         Task<List<string>> GetTags();
         Task<List<string>> GetCategories();
-        Task<int> DeleteTransactions(List<int> keys);
-        Task<bool> DeleteTransaction(int key);
+        Task<int> DeleteTransactions(List<long> keys);
+        Task<bool> DeleteTransaction(long key);
         Task<int> DeleteTransactions(List<TransactionFilter> filters);
     }
 }
